editgroups.cgi: code style, use hash lookup instead of grep

+++ editgroups.cgi
@@ -1,6 +1,4 @@
 #!/usr/bin/perl -wT
-# -*- Mode: perl; indent-tabs-mode: nil -*-
-#
 # The contents of this file are subject to the Mozilla Public
 # License Version 1.1 (the "License"); you may not use this file
 # except in compliance with the License. You may obtain a copy of
@@ -55,14 +53,14 @@ my $token  = $cgi->param('token');
 # actually a valid group ID. If all tests are successful, the
 # trimmed group ID is returned.
 
-sub CheckGroupID {
+sub CheckGroupID
+{
     my ($group_id) = @_;
     $group_id = trim($group_id || 0);
     ThrowUserError("group_not_specified") unless $group_id;
     (detaint_natural($group_id)
-      && Bugzilla->dbh->selectrow_array("SELECT id FROM groups WHERE id = ?",
-                                        undef, $group_id))
-      || ThrowUserError("invalid_group_ID");
+        && Bugzilla->dbh->selectrow_array("SELECT id FROM groups WHERE id = ?", undef, $group_id))
+        || ThrowUserError("invalid_group_ID");
     return $group_id;
 }
 
@@ -71,7 +69,8 @@ sub CheckGroupID {
 # if none is given, as expected). The trimmed regular expression
 # is returned.
 
-sub CheckGroupRegexp {
+sub CheckGroupRegexp
+{
     my ($regexp) = @_;
     $regexp = trim($regexp || '');
     trick_taint($regexp);
@@ -80,18 +79,20 @@ sub CheckGroupRegexp {
 }
 
 # A helper for displaying the edit.html.tmpl template.
-sub get_current_and_available {
+sub get_current_and_available
+{
     my ($group, $vars) = @_;
 
     my @all_groups         = Bugzilla::Group->get_all;
-    my @members_current    = @{$group->grant_direct(GROUP_MEMBERSHIP)};
-    my @member_of_current  = @{$group->granted_by_direct(GROUP_MEMBERSHIP)};
-    my @bless_from_current = @{$group->grant_direct(GROUP_BLESS)};
-    my @bless_to_current   = @{$group->granted_by_direct(GROUP_BLESS)};
-    my (@visible_from_current, @visible_to_me_current);
-    if (Bugzilla->params->{'usevisibilitygroups'}) {
-        @visible_from_current  = @{$group->grant_direct(GROUP_VISIBLE)};
-        @visible_to_me_current = @{$group->granted_by_direct(GROUP_VISIBLE)};
+    my $members_current    = { map { $_->id => $_ } @{$group->grant_direct(GROUP_MEMBERSHIP)} };
+    my $member_of_current  = { map { $_->id => $_ } @{$group->granted_by_direct(GROUP_MEMBERSHIP)} };
+    my $bless_from_current = { map { $_->id => $_ } @{$group->grant_direct(GROUP_BLESS)} };
+    my $bless_to_current   = { map { $_->id => $_ } @{$group->granted_by_direct(GROUP_BLESS)} };
+    my ($visible_from_current, $visible_to_me_current);
+    if (Bugzilla->params->{usevisibilitygroups})
+    {
+        $visible_from_current  = { map { $_->id => $_ } @{$group->grant_direct(GROUP_VISIBLE)} };
+        $visible_to_me_current = { map { $_->id => $_ } @{$group->granted_by_direct(GROUP_VISIBLE)} };
     }
 
     # Figure out what groups are not currently a member of this group,
@@ -99,16 +100,15 @@ sub get_current_and_available {
     my (@members_available, @member_of_available,
         @bless_from_available, @bless_to_available,
         @visible_from_available, @visible_to_me_available);
-    foreach my $group_option (@all_groups) {
-        if (Bugzilla->params->{'usevisibilitygroups'}) {
-            push(@visible_from_available, $group_option)
-                if !grep($_->id == $group_option->id, @visible_from_current);
-            push(@visible_to_me_available, $group_option)
-                if !grep($_->id == $group_option->id, @visible_to_me_current);
+    foreach my $group_option (@all_groups)
+    {
+        if (Bugzilla->params->{usevisibilitygroups})
+        {
+            push @visible_from_available, $group_option if !$visible_from_current->{$group_option->id};
+            push @visible_to_me_available, $group_option if !$visible_to_me_current->{$group_option->id};
         }
 
-        push(@bless_from_available, $group_option)
-            if !grep($_->id == $group_option->id, @bless_from_current);
+        push @bless_from_available, $group_option if !$bless_from_current->{$group_option->id};
 
         # The group itself should never show up in the membership lists,
         # and should show up in only one of the bless lists (otherwise
@@ -116,29 +116,27 @@ sub get_current_and_available {
         # database unique constraint error).
         next if $group_option->id == $group->id;
 
-        push(@members_available, $group_option)
-            if !grep($_->id == $group_option->id, @members_current);
-        push(@member_of_available, $group_option)
-            if !grep($_->id == $group_option->id, @member_of_current);
-        push(@bless_to_available, $group_option)
-           if !grep($_->id == $group_option->id, @bless_to_current);
+        push @members_available, $group_option if !$members_current->{$group_option->id};
+        push @member_of_available, $group_option if !$member_of_current->{$group_option->id};
+        push @bless_to_available, $group_option if !$bless_to_current->{$group_option->id};
     }
 
-    $vars->{'members_current'}     = \@members_current;
-    $vars->{'members_available'}   = \@members_available;
-    $vars->{'member_of_current'}   = \@member_of_current;
-    $vars->{'member_of_available'} = \@member_of_available;
+    $vars->{members_current}     = [ values %$members_current ];
+    $vars->{members_available}   = \@members_available;
+    $vars->{member_of_current}   = [ values %$member_of_current ];
+    $vars->{member_of_available} = \@member_of_available;
 
-    $vars->{'bless_from_current'}   = \@bless_from_current;
-    $vars->{'bless_from_available'} = \@bless_from_available;
-    $vars->{'bless_to_current'}     = \@bless_to_current;
-    $vars->{'bless_to_available'}   = \@bless_to_available;
+    $vars->{bless_from_current}   = [ values %$bless_from_current ];
+    $vars->{bless_from_available} = \@bless_from_available;
+    $vars->{bless_to_current}     = [ values %$bless_to_current ];
+    $vars->{bless_to_available}   = \@bless_to_available;
 
-    if (Bugzilla->params->{'usevisibilitygroups'}) {
-        $vars->{'visible_from_current'}    = \@visible_from_current;
-        $vars->{'visible_from_available'}  = \@visible_from_available;
-        $vars->{'visible_to_me_current'}   = \@visible_to_me_current;
-        $vars->{'visible_to_me_available'} = \@visible_to_me_available;
+    if (Bugzilla->params->{usevisibilitygroups})
+    {
+        $vars->{visible_from_current}    = [ values %$visible_from_current ];
+        $vars->{visible_from_available}  = \@visible_from_available;
+        $vars->{visible_to_me_current}   = [ values %$visible_to_me_current ];
+        $vars->{visible_to_me_available} = \@visible_to_me_available;
     }
 }
 
@@ -156,29 +154,29 @@ unless ($action)
 
 # All other actions are protected by the 'creategroups' permission.
 
-$vars->{allow_edit}
-  || ThrowUserError("auth_failure", {group  => "creategroups",
-                                     action => "edit",
-                                     object => "groups"});
+$vars->{allow_edit} || ThrowUserError("auth_failure", {
+    group  => "creategroups",
+    action => "edit",
+    object => "groups",
+});
 
 #
 # action='changeform' -> present form for altering an existing group
 #
 # (next action will be 'postchanges')
 #
-
-if ($action eq 'changeform') {
+if ($action eq 'changeform')
+{
     # Check that an existing group ID is given
     my $group_id = CheckGroupID($cgi->param('group'));
     my $group = new Bugzilla::Group($group_id);
 
     get_current_and_available($group, $vars);
-    $vars->{'group'} = $group;
-    $vars->{'token'}       = issue_session_token('edit_group');
+    $vars->{group} = $group;
+    $vars->{token} = issue_session_token('edit_group');
 
     $template->process("admin/groups/edit.html.tmpl", $vars)
-      || ThrowTemplateError($template->error());
-
+        || ThrowTemplateError($template->error());
     exit;
 }
 
@@ -187,22 +185,19 @@ if ($action eq 'changeform') {
 #
 # (next action will be 'new')
 #
-
-if ($action eq 'add') {
-    $vars->{'token'} = issue_session_token('add_group');
+if ($action eq 'add')
+{
+    $vars->{token} = issue_session_token('add_group');
     $template->process("admin/groups/create.html.tmpl", $vars)
-      || ThrowTemplateError($template->error());
-    
+        || ThrowTemplateError($template->error());
     exit;
 }
-
-
 
 #
 # action='new' -> add group entered in the 'action=add' screen
 #
-
-if ($action eq 'new') {
+if ($action eq 'new')
+{
     check_token_data($token, 'add_group');
     my $group = Bugzilla::Group->create({
         name        => scalar $cgi->param('name'),
@@ -214,24 +209,26 @@ if ($action eq 'new') {
     });
 
     # Permit all existing products to use the new group.
-    if ($cgi->param('insertnew')) {
-        $dbh->do('INSERT INTO group_control_map
-                  (group_id, product_id, membercontrol, othercontrol)
-                  SELECT ?, products.id, ?, ? FROM products',
-                  undef, ($group->id, CONTROLMAPSHOWN, CONTROLMAPNA));
+    if ($cgi->param('insertnew'))
+    {
+        $dbh->do(
+            'INSERT INTO group_control_map (group_id, product_id, membercontrol, othercontrol)'.
+            ' SELECT ?, products.id, ?, ? FROM products',
+            undef, $group->id, CONTROLMAPSHOWN, CONTROLMAPNA
+        );
     }
 
     Bugzilla::Hook::process('editgroups-post_create', { group => $group });
 
     delete_token($token);
 
-    $vars->{'message'} = 'group_created';
-    $vars->{'group'} = $group;
+    $vars->{message} = 'group_created';
+    $vars->{group} = $group;
     get_current_and_available($group, $vars);
-    $vars->{'token'} = issue_session_token('edit_group');
+    $vars->{token} = issue_session_token('edit_group');
 
     $template->process("admin/groups/edit.html.tmpl", $vars)
-      || ThrowTemplateError($template->error());
+        || ThrowTemplateError($template->error());
     exit;
 }
 
@@ -240,34 +237,32 @@ if ($action eq 'new') {
 #
 # (next action would be 'delete')
 #
-
-if ($action eq 'del') {
+if ($action eq 'del')
+{
     # Check that an existing group ID is given
     my $group = Bugzilla::Group->check({ id => $cgi->param('group') });
     $group->check_remove({ test_only => 1 });
-    $vars->{'shared_queries'} =
-        $dbh->selectrow_array('SELECT COUNT(*)
-                                 FROM namedquery_group_map
-                                WHERE group_id = ?', undef, $group->id);
+    $vars->{shared_queries} = $dbh->selectrow_array(
+        'SELECT COUNT(*) FROM namedquery_group_map WHERE group_id = ?', undef, $group->id
+    );
 
-    $vars->{'group'} = $group;
-    $vars->{'token'} = issue_session_token('delete_group');
+    $vars->{group} = $group;
+    $vars->{token} = issue_session_token('delete_group');
 
     $template->process("admin/groups/delete.html.tmpl", $vars)
-      || ThrowTemplateError($template->error());
-    
+        || ThrowTemplateError($template->error());
     exit;
 }
 
 #
 # action='delete' -> really delete the group
 #
-
-if ($action eq 'delete') {
+if ($action eq 'delete')
+{
     check_token_data($token, 'delete_group');
     # Check that an existing group ID is given
     my $group = Bugzilla::Group->check({ id => $cgi->param('group') });
-    $vars->{'name'} = $group->name;
+    $vars->{name} = $group->name;
     $group->remove_from_db({
         remove_from_users => scalar $cgi->param('removeusers'),
         remove_from_bugs  => scalar $cgi->param('removebugs'),
@@ -279,7 +274,7 @@ if ($action eq 'delete') {
     Bugzilla::Hook::process('editgroups-post_delete', { group => $group });
     Bugzilla::Views::refresh_some_views();
 
-    $vars->{'message'} = 'group_deleted';
+    $vars->{message} = 'group_deleted';
     ListGroups($vars);
     exit;
 }
@@ -287,8 +282,8 @@ if ($action eq 'delete') {
 #
 # action='postchanges' -> update the groups
 #
-
-if ($action eq 'postchanges') {
+if ($action eq 'postchanges')
+{
     check_token_data($token, 'edit_group');
     my $changes = doGroupChanges();
 
@@ -299,27 +294,29 @@ if ($action eq 'postchanges') {
 
     my $group = new Bugzilla::Group($cgi->param('group_id'));
     get_current_and_available($group, $vars);
-    $vars->{'message'} = 'group_updated';
-    $vars->{'group'}   = $group;
-    $vars->{'changes'} = $changes;
-    $vars->{'token'} = issue_session_token('edit_group');
+    $vars->{message} = 'group_updated';
+    $vars->{group}   = $group;
+    $vars->{changes} = $changes;
+    $vars->{token} = issue_session_token('edit_group');
 
     $template->process("admin/groups/edit.html.tmpl", $vars)
-      || ThrowTemplateError($template->error());
+        || ThrowTemplateError($template->error());
     exit;
 }
 
-if ($action eq 'confirm_remove') {
+if ($action eq 'confirm_remove')
+{
     my $group = new Bugzilla::Group(CheckGroupID($cgi->param('group_id')));
-    $vars->{'group'} = $group;
-    $vars->{'regexp'} = CheckGroupRegexp($cgi->param('regexp'));
-    $vars->{'token'} = issue_session_token('remove_group_members');
+    $vars->{group} = $group;
+    $vars->{regexp} = CheckGroupRegexp($cgi->param('regexp'));
+    $vars->{token} = issue_session_token('remove_group_members');
     $template->process('admin/groups/confirm-remove.html.tmpl', $vars)
         || ThrowTemplateError($template->error());
     exit;
 }
 
-if ($action eq 'remove_regexp') {
+if ($action eq 'remove_regexp')
+{
     check_token_data($token, 'remove_group_members');
     # remove all explicit users from the group with
     # gid = $cgi->param('group') that match the regular expression
@@ -331,34 +328,33 @@ if ($action eq 'remove_regexp') {
     $dbh->bz_start_transaction();
 
     my $users = $group->members_direct();
-    my $sth_delete = $dbh->prepare(
-        "DELETE FROM user_group_map
-           WHERE user_id = ? AND isbless = 0 AND group_id = ?");
+    my $sth_delete = $dbh->prepare("DELETE FROM user_group_map WHERE user_id = ? AND isbless = 0 AND group_id = ?");
 
     my @deleted;
-    foreach my $member (@$users) {
-        if ($regexp eq '' || $member->login =~ m/$regexp/i) {
+    foreach my $member (@$users)
+    {
+        if ($regexp eq '' || $member->login =~ m/$regexp/i)
+        {
             $sth_delete->execute($member->id, $group->id);
-            push(@deleted, $member);
+            push @deleted, $member;
         }
     }
     $dbh->bz_commit_transaction();
 
-    $vars->{'users'}  = \@deleted;
-    $vars->{'regexp'} = $regexp;
+    $vars->{users}  = \@deleted;
+    $vars->{regexp} = $regexp;
 
     Bugzilla::Hook::process('editgroups-post_remove_regexp', { deleted => \@deleted });
     Bugzilla::Views::refresh_some_views();
 
     delete_token($token);
 
-    $vars->{'message'} = 'group_membership_removed';
-    $vars->{'group'} = $group->name;
+    $vars->{message} = 'group_membership_removed';
+    $vars->{group} = $group->name;
 
     ListGroups($vars);
     exit;
 }
-
 
 #
 # No valid action found
@@ -366,9 +362,9 @@ if ($action eq 'remove_regexp') {
 
 ThrowCodeError("action_unrecognized", $vars);
 
-
 # Helper sub to handle the making of changes to a group
-sub doGroupChanges {
+sub doGroupChanges
+{
     my $cgi = Bugzilla->cgi;
     my $dbh = Bugzilla->dbh;
 
@@ -377,39 +373,43 @@ sub doGroupChanges {
     # Check that the given group ID is valid and make a Group.
     my $group = new Bugzilla::Group(CheckGroupID($cgi->param('group_id')));
 
-    if (defined $cgi->param('regexp')) {
+    if (defined $cgi->param('regexp'))
+    {
         $group->set_user_regexp($cgi->param('regexp'));
     }
 
-    if ($group->is_bug_group) {
-        if (defined $cgi->param('name')) {
+    if ($group->is_bug_group)
+    {
+        if (defined $cgi->param('name'))
+        {
             $group->set_name($cgi->param('name'));
         }
-        if (defined $cgi->param('desc')) {
+        if (defined $cgi->param('desc'))
+        {
             $group->set_description($cgi->param('desc'));
         }
         # Only set isactive if we came from the right form.
-        if (defined $cgi->param('regexp')) {
+        if (defined $cgi->param('regexp'))
+        {
             $group->set_is_active($cgi->param('isactive'));
         }
     }
 
-    if (defined $cgi->param('icon_url')) {
+    if (defined $cgi->param('icon_url'))
+    {
         $group->set_icon_url($cgi->param('icon_url'));
     }
 
     my $changes = $group->update();
 
-    my $sth_insert = $dbh->prepare('INSERT INTO group_group_map
-                                    (member_id, grantor_id, grant_type)
-                                    VALUES (?, ?, ?)');
+    my $sth_insert = $dbh->prepare(
+        'INSERT INTO group_group_map (member_id, grantor_id, grant_type) VALUES (?, ?, ?)'
+    );
+    my $sth_delete = $dbh->prepare(
+        'DELETE FROM group_group_map WHERE member_id = ? AND grantor_id = ? AND grant_type = ?'
+    );
 
-    my $sth_delete = $dbh->prepare('DELETE FROM group_group_map
-                                     WHERE member_id = ?
-                                           AND grantor_id = ?
-                                           AND grant_type = ?');
-
-    # First item is the type, second is whether or not it's "reverse" 
+    # First item is the type, second is whether or not it's "reverse"
     # (granted_by) (see _do_add for more explanation).
     my %fields = (
         members       => [GROUP_MEMBERSHIP, 0],
@@ -419,38 +419,41 @@ sub doGroupChanges {
         bless_to      => [GROUP_BLESS, 1],
         visible_to_me => [GROUP_VISIBLE, 1]
     );
-    while (my ($field, $data) = each %fields) {
-        _do_add($group, $changes, $sth_insert, "${field}_add", 
-                $data->[0], $data->[1]);
-        _do_remove($group, $changes, $sth_delete, "${field}_remove",
-                   $data->[0], $data->[1]);
+    while (my ($field, $data) = each %fields)
+    {
+        _do_add($group, $changes, $sth_insert, "${field}_add", $data->[0], $data->[1]);
+        _do_remove($group, $changes, $sth_delete, "${field}_remove", $data->[0], $data->[1]);
     }
 
     $dbh->bz_commit_transaction();
     return $changes;
 }
 
-sub _do_add {
+sub _do_add
+{
     my ($group, $changes, $sth_insert, $field, $type, $reverse) = @_;
     my $cgi = Bugzilla->cgi;
 
     my $current;
     # $reverse means we're doing a granted_by--that is, somebody else
     # is granting us something.
-    if ($reverse) {
+    if ($reverse)
+    {
         $current = $group->granted_by_direct($type);
     }
-    else {
+    else
+    {
         $current = $group->grant_direct($type);
     }
 
     my $add_items = Bugzilla::Group->new_from_list([$cgi->param($field)]);
 
-    foreach my $add (@$add_items) {
+    foreach my $add (@$add_items)
+    {
         next if grep($_->id == $add->id, @$current);
 
         $changes->{$field} ||= [];
-        push(@{$changes->{$field}}, $add->name);
+        push @{$changes->{$field}}, $add->name;
         # They go this direction for a normal "This group is granting
         # $add something."
         my @ids = ($add->id, $group->id);
@@ -461,12 +464,14 @@ sub _do_add {
     }
 }
 
-sub _do_remove {
+sub _do_remove
+{
     my ($group, $changes, $sth_delete, $field, $type, $reverse) = @_;
     my $cgi = Bugzilla->cgi;
     my $remove_items = Bugzilla::Group->new_from_list([$cgi->param($field)]);
 
-    foreach my $remove (@$remove_items) {
+    foreach my $remove (@$remove_items)
+    {
         my @ids = ($remove->id, $group->id);
         # See _do_add for an explanation of $reverse
         @ids = reverse @ids if $reverse;
@@ -474,7 +479,7 @@ sub _do_remove {
         # don't need to do any checks.
         $sth_delete->execute(@ids, $type);
         $changes->{$field} ||= [];
-        push(@{$changes->{$field}}, $remove->name);
+        push @{$changes->{$field}}, $remove->name;
     }
 }
